fix: extension range protocol wrong nodes deletion
There are rare occasions when the left worker creates a page that was previously
elided by the right worker. In this case, this fix ensures that the page
is properly updated without deleting it (which was happening before).

The reason this happens is that the leaf and branch stages, while handling merges,
update the cutoff of each node. The cutoff updated after the deletion of a node
can also be used as a new high range sent to the left worker.
Thus, the left worker thinks that everything between the last modified page sent
by the right worker and the new high range has free keys to be used as separators,
and those separators could have previously been eliminated by the right worker.